refactor: 클래스 명 파스칼 케이스로 수정
