+++ src/main/java/com/somemore/center/service/RegisterCenterService.java
@@ -2,6 +2,7 @@ package com.somemore.center.service;
 
 import com.somemore.center.domain.Center_NEW;
 import com.somemore.center.repository.CenterRepository;
+import com.somemore.center.usecase.RegisterCenterUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,7 +12,7 @@ import java.util.UUID;
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class registerCenterService implements registerCenterUseCase {
+public class RegisterCenterService implements RegisterCenterUseCase {
 
     private final CenterRepository centerRepository;
