+++ src/main/java/com/somemore/center/usecase/RegisterCenterUseCase.java
@@ -1,10 +1,10 @@
-package com.somemore.center.service;
+package com.somemore.center.usecase;
 
 import com.somemore.center.domain.Center_NEW;
 
 import java.util.UUID;
 
-public interface registerCenterUseCase {
+public interface RegisterCenterUseCase {
 
     Center_NEW register(UUID userId);
 }
