+++ src/main/java/com/somemore/user/service/RegisterUserUseCase.java
@@ -9,7 +9,6 @@ import com.somemore.user.dto.UserAuthInfo;
 import com.somemore.user.repository.user.UserRepository;
 import com.somemore.user.repository.usercommonattribute.UserCommonAttributeRepository;
 import com.somemore.user.repository.usersetting.UserSettingRepository;
-import com.somemore.user.usecase.RegisterUserUseCase;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -21,7 +20,7 @@ import java.util.UUID;
 @Slf4j
 @RequiredArgsConstructor
 @Transactional
-public class RegisterUserService implements RegisterUserUseCase {
+public class RegisterUserUseCase implements com.somemore.user.usecase.RegisterUserUseCase {
 
     private final UserRepository userRepository;
     private final UserSettingRepository userSettingRepository;
