+++ src/main/java/com/somemore/volunteer/service/RegisterVolunteerService.java
@@ -2,6 +2,7 @@ package com.somemore.volunteer.service;
 
 import com.somemore.volunteer.domain.Volunteer_NEW;
 import com.somemore.volunteer.repository.VolunteerRepository;
+import com.somemore.volunteer.usecase.RegisterVolunteerUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,7 +12,7 @@ import java.util.UUID;
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class registerVolunteerService implements registerVolunteerUseCase {
+public class RegisterVolunteerService implements RegisterVolunteerUseCase {
 
     private final VolunteerRepository volunteerRepository;
