+++ src/main/java/com/somemore/volunteer/usecase/RegisterVolunteerUseCase.java
@@ -1,10 +1,10 @@
-package com.somemore.volunteer.service;
+package com.somemore.volunteer.usecase;
 
 import com.somemore.volunteer.domain.Volunteer_NEW;
 
 import java.util.UUID;
 
-public interface registerVolunteerUseCase {
+public interface RegisterVolunteerUseCase {
 
     Volunteer_NEW register(UUID userId);
 }
